did some restructuring and renaming
