+++ src/api/manga_tags.go
@@ -1,0 +1,2 @@
+// this is to make calls to get manga tags etc.
+package api
