+++ src/api/manga_volumes_chapters.go
@@ -1,0 +1,2 @@
+// this is to get the a mangas series volumes and chapters data
+package api
